src/requirements.txt
paho-mqtt==1.5.0
RPi.GPIO==0.7.0
